You decide to join a very large known gym chain. When you arrive, you go to the reception. You find the owner arguing with a member who wants to attend a session he has not registered for. Your turn finally arrives, the owner apologizes for letting you wait so long. He explains that his employee, in charge of registrations, made a mistake in choosing the member's session.
You are surprised that such a recognized gymnasium still uses the cards to manage its large number of clients. So you're talking to the owner to  fix this problem. Interested in your knowledge, he asks you to find him a solution, in return he will offer you a free 3-months inscription.